rename exceptions.py -> errors.py
