+++ test/test_errors.py
@@ -8,7 +8,7 @@ import pytest
 from grpc import StatusCode
 
 from nameko_grpc.constants import Cardinality
-from nameko_grpc.exceptions import GrpcError
+from nameko_grpc.errors import GrpcError
 
 
 @pytest.mark.equivalence
